+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 * Add method to get total count of backlinks for a row.
   PR [#2672](https://github.com/realm/realm-core/pull/2672).
+* Add try_remove_dir() and try_remove_dir_recursive() functions.
 
 -----------